Accessible legend
Added an accessible dynamic legend to the top right corner of the map that has a h2 heading indicating 'Map legend layers' not displayed to a sighted user. Using the title a h3 tag is added to indicate the name of the map layer. To ensure the page heading structure is met a h1 tag was added to the map, also hidden to sighted users. Tested on a Mac with Voiceover, aXe, WAVE, and ChromeVox.